windows_pc_code/requirements.txt
pyserial
pynput